feat: 识别智能化——图片预检+字段校验+营养素自洽+包装容量优先+反馈闭环
基于开源调研融入学术验证方案（两步 prompt + 营养素自洽校验 + 包装容量优先），
提升拍照识别准确率，减少出错概率。

批次1 图片预检+字段校验+营养素自洽：
- 新建 recognition_validator.dart：字段合理性校验（dishName/confidence/weight/区间）
  + 营养素自洽校验（4*protein+9*fat+4*carb ≈ calories，误差±10%）
- recognize_controller 集成校验：字段不合理→重试1次；营养素不自洽→自动修正
- 修复 image_quality_checker.dart 类型错误（num→double）
- 20个校验器单元测试

批次2 prompt v1.6 包装容量优先：
- 包装食品必须读取包装标签净含量（weight_source=package_label），不靠视觉估算
- 营养素自洽约束：AI 用 4p+9f+4c 反算 calories，偏差<5%
- VisionRecognitionResult 新增 weightSource 字段（向后兼容旧 prompt）

批次3 反馈闭环回流 aliasesJson：
- FoodItemRepository 新增 addAlias 方法（事务+归一化去重）
- _showFeedbackDialog 加别名回流：用户纠正菜名后，AI错误名作为正确菜的别名
- 下次AI识别返回错误名时，findByNameOrAlias 命中别名，直接返回正确菜
- 6个 addAlias 单元测试

验证：flutter analyze No issues + flutter test 268 passed/3 skipped/0 failed
未发布 release（等用户确认）

+++ lib/core/util/image_quality_checker.dart
@@ -62,7 +62,7 @@ class ImageQualityChecker {
         final left = img.getLuminance(gray.getPixel(x - 1, y));
         final right = img.getLuminance(gray.getPixel(x + 1, y));
         final lap = up + down + left + right - 4 * center;
-        laplacianValues.add(lap);
+        laplacianValues.add(lap.toDouble());
       }
     }
